borrowing fully functioning yey.

+++ src/main/java/com/example/demo/borrowings/BorrowingController.java
@@ -35,7 +35,7 @@ public class BorrowingController {
     }
 
     @PostMapping("/api/borrowings")
-    public Long createBorrowing(@RequestBody BorrowingDto borrowingDto){
+    public int createBorrowing(@RequestBody BorrowingDto borrowingDto){
         return borrowingService.createBorrowing(borrowingDto);
     }
